+++ examples-gallery/plot_vyasarayani.py
@@ -109,6 +109,10 @@ print(identified_msg)
 print(divider)
 
 # %%
+# Plot constraint violations.
+prob.plot_constraint_violations(solution)
+
+# %%
 # Simulate with the identified parameter.
 y_sim = odeint(eval_f, y0, time, args=(p_sol,))
 y1_sim = y_sim[:, 0]
@@ -133,6 +137,8 @@ axes_y1[1].set_title('Initial Guess Constraint Violations')
 axes_y1[1].plot(prob.con(initial_guess)[:num_nodes - 1])
 axes_y1[2].set_title('Solution Constraint Violations')
 axes_y1[2].plot(prob.con(solution)[:num_nodes - 1])
+
+# sphinx_gallery_thumbnail_number = 2
 
 # %%
 fig_y2, axes_y2 = plt.subplots(3, 1, layout='constrained')